Some language fix, and Log screen rotate fix
git-svn-id: https://svn.code.sf.net/p/geokretylog/code@121 ff536b33-f402-454f-b109-af6ef765b36a

+++ trunk/src/pl/nkg/geokrety/data/GeoKretLog.java
@@ -100,7 +100,7 @@ public class GeoKretLog {
 			latlon = savedInstanceState.getString(LATLON);
 			wpt = savedInstanceState.getString(WAYPOINT);
 
-			accoundID = savedInstanceState.getInt(ACCOUNT_ID);
+			accoundID = savedInstanceState.getLong(ACCOUNT_ID);
 		}
 	}
